Add a function to get waveform data from T8 API.
Add a function to get waveform data from T8 API given a host, machine,
point, pmode, timestamp, user, and password. The function makes an HTTP
GET request to the T8 API and returns the raw waveform data and sample
rate.

currently, the function does not decode the waveform data, which is
encoded as a base64 zlib compressed string. This will be done in a
future commit.