Make the DataFetcher interface truly generic
The `DataFetcher` interface was not entirely generic, as its methods
depended on structures with URL parameters. This prevented the creation
of other data fetchers, for example, to read from a gRPC server.
Therefore, these structures have been replaced with others that store
the information to identify a record stored in a T8 given its machine,
point, processing mode, and datetime. The remaining parameters for the
HTTP case (the host, user and password) are passed through the
`HttpDataFetcher` structure, paving the way for implementing new data
fetchers.